I dont even remember what bugs we fixed but more stuff works

+++ src/edu/mit/compilers/ir/expression/literal/IRIntLiteral.java
@@ -2,14 +2,15 @@ package edu.mit.compilers.ir.expression.literal;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.math.BigInteger;
 
 import edu.mit.compilers.ir.IRNode;
 import edu.mit.compilers.ir.IRType;
 import edu.mit.compilers.ir.expression.IRExpression;
 
-public class IRIntLiteral extends IRLiteral<Integer> {
+public class IRIntLiteral extends IRLiteral<BigInteger> {
 
-	public IRIntLiteral(Integer value) {
+	public IRIntLiteral(BigInteger value) {
 		super(value);
 		expressionType = IRExpression.ExpressionType.INT_LITERAL;
 	}
